update MemHeader to hold the flags
The MemHeader structure is updated to hold the memory attribute flags.
This change is triggered by the need to hold the coreid in the flags from
the libdce perpective.
In future flags can be used to hold more attributes.
The least significant 4 bits is assumed to hold the coreid.
This information is noted needed by the remoteproc. Its only for
book keeping and the structure sizes on MPU and remote procs must match.

dependent MPU Patch:
http://review.omapzoom.org/#/c/34863

Signed-off-by: Pradeep Venkatasubbarao <pradeepv@ti.com>